pyproject.toml
[tool.poetry]
name = "valkey-setup"
version = "0.1.0"
description = ""
authors = ["Declan Joseph <declan.obu@outlook.com>"]
readme = "README.md"

packages = [{ include = "valkey_setup", from = "src" }]

[tool.poetry.dependencies]
python = "^3.13"
poetry-core = "^2.1.3"
typer = "^0.20.1"
pyyaml = "^6.0.3"
rich = "^14.2.0"
pydantic = "^2.12.5"
sh = "^2.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
